+++ packages/govuk-frontend/src/govuk/components/radios/template.test.js
@@ -262,6 +262,15 @@ describe('Radios', () => {
         const $input = $('.govuk-radios__input').first()
         expect($input.attr('data-aria-controls')).toBeFalsy()
       })
+
+      // Indentation in nunjucks can mutate the value of textareas, since
+      // textarea value is defined between the html tags
+      it('does not add space to the input value of textareas inside conditionals', () => {
+        const $ = render('radios', examples['textarea in conditional'])
+
+        const $textarea = $('#conditional-textarea')
+        expect($textarea.text()).toBe('test\n')
+      })
     })
 
     it('render divider', () => {